consolidate api, improve how examples are included in docs
